+++ projects/demo/src/app/counter/counter.component.ts
@@ -3,7 +3,9 @@ import {
   ChangeDetectionStrategy,
   Component,
   ElementRef,
+  Signal,
   ViewChild,
+  computed,
 } from '@angular/core';
 import { Observable } from 'rxjs';
 import {