+++ projects/ngx-state-service/CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2024-02-??
+
+WORKING HERE
+
+### Added
+
+- New `StateService.valueSignal` property propagates changes of state as signal.
+
+### Changed
+
+- Option `isDeep` removed from `StateSettingOptions`.
+- `set` uses `Partial<T>` type for `statusUpdate` parameter now.
+- `mut` uses `Partial<T>` type for `props` parameter now.
+
+### Fixed
+
+- Fixed generics in signatures of `set`, `setDeep`, `mut` and `mutDeep`.
+
 ## [0.4.0] - 2024-02-15
 
 ### Added